AScontrol-classes {arules}	R Documentation

Classes AScontrol, APcontrol, ECcontrol — Specifying the control Argument of apriori() and eclat()

Description

The AScontrol class holds the algorithmic parameters for the used mining algorithms. APcontrol and ECcontrol directly extend AScontrol with additional slots for parameters only suitable for the algorithms Apriori (APcontrol) and Eclat (ECcontrol).

Objects from the Class

A suitable default control object will be automatically created by the apriori or the eclat function. By specifying a named list (names equal to slots) as control argument for the apriori or the eclat function, default values can be replaced by the values in the list. Objects can also be created by calls of the form new("APcontrol", ...) or new("ECcontrol", ...).

Slots

Common slots defined in AScontrol:

sort:

an integer scalar indicating how to sort items with respect to their frequency: (default: 2)

1:

ascending

-1:

descending

0:

do not sort

2:

ascending

-2:

descending with respect to transaction size sum

verbose:

a logical indicating whether to report progress

Additional slots for Apriori in APcontrol:

filter:

a numeric scalar indicating how to filter unused items from transactions (default: 0.1)

=0:

do not filter items with respect to. usage in sets

<0:

fraction of removed items for filtering

>0:

take execution times ratio into account

tree:

a logical indicating whether to organize transactions as a prefix tree (default: TRUE)

heap:

a logical indicating whether to use heapsort instead of quicksort to sort the transactions (default: TRUE)

memopt:

a logical indicating whether to minimize memory usage instead of maximize speed (default: FALSE)

load:

a logical indicating whether to load transactions into memory (default: TRUE)

Additional slots for Eclat in ECcontrol:

sparse:

a numeric value for the threshold for sparse representation (default: 7)

Methods

coerce

signature(from = "NULL", to = "APcontrol")

coerce

signature(from = "list", to = "APcontrol")

coerce

signature(from = "NULL", to = "ECcontrol")

coerce

signature(from = "list", to = "ECcontrol")

Author(s)

Michael Hahsler and Bettina Gruen

References

Christian Borgelt (2004) Apriori — Finding Association Rules/Hyperedges with the Apriori Algorithm. www.borgelt.net/apriori.html

See Also

apriori, eclat

[Package arules version 1.6-1 Index]