Epub {arules}	R Documentation

Epub Data Set

Description

The Epub data set contains the download history of documents from the electronic publication platform of the Vienna University of Economics and Business Administration. The data was recorded between Jan 2003 and Dec 2008.

Usage

data(Epub)

Format

Object of class transactions with 15729 transactions and 936 items. Item labels are document IDs of the from "doc\_11d". Session IDs and time stamps for transactions are also provided.

Author(s)

Michael Hahsler

Source

Provided by Michael Hahsler from ePub-WU at http://epub.wu-wien.ac.at.

[Package arules version 1.6-1 Index]