combine {arules}	R Documentation

Combining Objects

Description

Provides the S4 methods to combine several objects based on itemMatrix into a single object.

Note, use union rather than c to combine several mined itemsets (or rules) into a single set.

Usage

## S4 method for signature 'itemMatrix'
c(x, ..., recursive = FALSE)

## S4 method for signature 'transactions'
c(x, ..., recursive = FALSE)

## S4 method for signature 'rules'
c(x, ..., recursive = FALSE)

## S4 method for signature 'itemsets'
c(x, ..., recursive = FALSE)

Arguments

x
first object.

...
further objects of the same class as x to be combined.

recursive
a logical. If recursive=TRUE, the function recursively descends through lists combining all their elements into a vector.

Value

An object of the same class as x.

Author(s)

Michael Hahsler

See Also

itemMatrix-class, transactions-class, rules-class, itemsets-class

Examples

data("Adult")

## combine transactions
a1 <- Adult[1:10]
a2 <- Adult[101:110]

aComb <- c(a1, a2)
summary(aComb)

## combine rules (can contain the same rule multiple times)
r1 <- apriori(Adult[1:1000])
r2 <- apriori(Adult[1001:2000])
rComb <- c(r1, r2) 
rComb

## union of rules (a set with only unique rules: same as unique(rComb))
rUnion <- union(r1,r2)
rUnion

[Package arules version 1.6-1 Index]