duplicated {arules}	R Documentation

Find Duplicated Elements

Description

Provides the generic function duplicated and the S4 methods for itemMatrix and associations. duplicated finds duplicated elements in an itemMatrix. It returns a logical vector indicating which elements are duplicates.

Note that duplicated can also be used to find transactions with identical items and identical rules and itemsets stored in rules and itemsets.

Usage

duplicated(x, incomparables = FALSE, ...)

Arguments

x
an object of class itemMatrix or associations.

...
further arguments (currently unused).

incomparables
argument currently unused.

Value

A logical vector indicating duplicated elements.

Author(s)

Michael Hahsler

See Also

unique, rules-class, itemsets-class, itemMatrix-class

Examples

data("Adult")

r1 <- apriori(Adult[1:1000], parameter = list(support = 0.5))
r2 <- apriori(Adult[1001:2000], parameter = list(support = 0.5))

## Note this creates a collection of rules from two sets of rules
r_comb <- c(r1, r2)
duplicated(r_comb)

[Package arules version 1.6-1 Index]