Methods for "[": Extraction or Subsetting arules Objects

Methods for "[", i.e., extraction or subsetting for arules objects.

Usage

# S4 method for class 'itemMatrix,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'transactions,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'tidLists,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'rules,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

# S4 method for class 'itemsets,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

Arguments

x

an object of class itemMatrix, transactions or associations.

i

select rows/sets using an integer vector containing row numbers or a logical vector.

j

select columns/items using an integer vector containing column numbers (i.e., item IDs), a logical vector or a vector of strings containing parts of item labels.

...

further arguments are ignored.

drop

ignored.

See also

Author

Michael Hahsler

Examples

data(Adult)
Adult
#> transactions in sparse format with
#>  48842 transactions (rows) and
#>  115 items (columns)

## select first 10 transactions
Adult[1:10]
#> transactions in sparse format with
#>  10 transactions (rows) and
#>  115 items (columns)

## select first 10 items for first 100 transactions
Adult[1:100, 1:10]
#> transactions in sparse format with
#>  100 transactions (rows) and
#>  10 items (columns)

## select the first 100 transactions for the items containing
## "income" or "age=Young" in their labels
Adult[1:100, c("income=small", "income=large", "age=Young")]
#> transactions in sparse format with
#>  100 transactions (rows) and
#>  3 items (columns)