is.maximal {arules}	R Documentation

Find Maximal Itemsets

Description

Provides the generic function and the S4 method is.maximal for finding maximal itemsets.

Usage

is.maximal(x,...)
## S4 method for signature 'itemMatrix'
is.maximal(x)

Arguments

x
the set of itemsets, rules or an itemMatrix object.

...
further arguments.

Details

An itemset is maximal in a set if no proper superset of the itemset is contained in the set (Zaki et al., 1997).

We define here maximal rules, as the rules generated by maximal itemsets.

Value

a logical vector with the same length as x indicating for each element in x if it is a maximal itemset.

Author(s)

Michael Hahsler

References

Mohammed J. Zaki, Srinivasan Parthasarathy, Mitsunori Ogihara, and Wei Li (1997). New algorithms for fast discovery of association rules. Technical Report 651, Computer Science Department, University of Rochester, Rochester, NY 14627.

See Also

is.superset, itemMatrix-class, itemsets-class

[Package arules version 1.6-1 Index]