length {arules}	R Documentation

Getting the Number of Elements

Description

S4 methods for length which return the number of elements of objects defined in the package arules.

Usage

## S4 method for signature 'rules'
length(x)

## S4 method for signature 'itemsets'
length(x)

## S4 method for signature 'tidLists'
length(x)

## S4 method for signature 'itemMatrix'
length(x)

Arguments

x
an object of class transactions, rules, itemsets, tidLists, or itemMatrix.

Details

For itemMatrix and transactions the length is defined as the number of rows (transactions) in the binary incidence matrix.

For sets of associations (rules, itemsets and associations in general) the length is defined as the number of elements in the set (i.e., the number of rules or itemsets).

For tidLists the length is the number of lists (one per item or itemset) in the object.

Value

An integer scalar giving the “length” of x.

Author(s)

Michael Hahsler

[Package arules version 1.6-1 Index]