size {arules}	R Documentation

Number of Items

Description

Provides the generic function size and S4 methods to get the size of each element from objects based on itemMatrix. For example, it is used to get a vector of transaction sizes (i.e., the number of present items (ones) per element (row) of the binary incidence matrix) from an object of class transactions).

Usage

size(x, ...)

Arguments

x
an object.

...
further (unused) arguments.

Value

size returns a numeric vector of length length(x). Each element is the size of the corresponding element (row in the matrix) in object x. For rules, size returns the sum of the number of elements in the LHS and the RHS.

Author(s)

Michael Hahsler

See Also

itemMatrix-class, transactions-class

Examples

data("Adult")
summary(size(Adult))

[Package arules version 1.6-1 Index]