Supporting Transactions

Find for each itemset in an associations object which transactions support (i.e., contains all items in the itemset) it. The information is returned as a tidLists object.

Usage

supportingTransactions(x, transactions, ...)

# S4 method for class 'associations'
supportingTransactions(x, transactions)

Arguments

x

a set of associations (itemsets, rules, etc.)

transactions

an object of class transactions used to mine the associations in x.

...

currently unused.

Value

An object of class tidLists containing one transaction ID list per association in x.

See also

Author

Michael Hahsler

Examples

data <- list(
  c("a", "b", "c"),
  c("a", "b"),
  c("a", "b", "d"),
  c("b", "e"),
  c("b", "c", "e"),
  c("a", "d", "e"),
  c("a", "c"),
  c("a", "b", "d"),
  c("c", "e"),
  c("a", "b", "d", "e")
)
data <- as(data, "transactions")

## mine itemsets
f <- eclat(data, parameter = list(support = .2, minlen = 3))
#> Eclat
#> 
#> parameter specification:
#>  tidLists support minlen maxlen            target  ext
#>     FALSE     0.2      3     10 frequent itemsets TRUE
#> 
#> algorithmic control:
#>  sparse sort verbose
#>       7   -2    TRUE
#> 
#> Absolute minimum support count: 2 
#> 
#> create itemset ... 
#> set transactions ...[5 item(s), 10 transaction(s)] done [0.00s].
#> sorting and recoding items ... [5 item(s)] done [0.00s].
#> creating bit matrix ... [5 row(s), 10 column(s)] done [0.00s].
#> writing  ... [2 set(s)] done [0.00s].
#> Creating S4 object  ... done [0.00s].
inspect(f)
#>     items     support count
#> [1] {a, d, e} 0.2     2    
#> [2] {a, b, d} 0.3     3    

## find supporting Transactions
st <- supportingTransactions(f, data)
st
#> tidLists in sparse format with
#>  2 items/itemsets (rows) and
#>  10 transactions (columns)

as(st, "list")
#> $`{a,d,e}`
#> [1]  6 10
#> 
#> $`{a,b,d}`
#> [1]  3  8 10
#>