Classes ASparameter, APparameter, ECparameter ---

The ASparameter class holds the mining parameters (e.g., minimum support) for the used mining algorithms. APparameter and ECparameter directly extend ASparameter with additional slots for parameters only suitable for the Apriori (APparameter) or the Eclat algorithms (ECparameter).

Objects from the Class

A suitable default parameter object will be automatically created by the apriori or the eclat function. By specifying a named list (names equal to slots) as parameter argument for the apriori or the eclat function, default values can be replaced by the values in the list. Objects can be created by calls of the form new("APparameter", ...) or new("ECparameter", ...).

Slots

Common slots defined in ASparameter:

support:

a numeric value for the minimal support of an item set (default: \(0.1\))

minlen:

an integer value for the minimal number of items per item set (default: 1 item)

maxlen:

an integer value for the maximal number of items per item set (default: 10 items)

target:

a character string indicating the type of association mined. One of

"frequent itemsets"

"maximally frequent itemsets"

"closed frequent itemsets"

"rules" (only available for Apriori)

"hyperedgesets" (only available for Apriori; see references for the definition of association hyperedgesets)

ext:

a logical indicating whether to produce extended information on quality measures (e.g., lhs.support) (default: FALSE)

Additional slots for Apriori in APparameter:

confidence:

a numeric value for the minimal confidence of rules/association hyperedges (default: \(0.8\)). For frequent itemsets it is set to NA.

smax:

a numeric value for the maximal support of itemsets/rules/hyperedgesets (default: 1)

arem:

a character string indicating the used additional rule evaluation measure (default: "none") given by one of

"none":

no additional evaluation measure

"diff":

absolute confidence difference

"quot":

difference of confidence quotient to 1

"aimp":

absolute difference of improvement to 1

"info":

information difference to prior

"chi2":

normalized \(\chi^2\) measure

aval:

a logical indicating whether to return the additional rule evaluation measure selected with arem.

minval:

a numeric value for the minimal value of additional evaluation measure selected with arem (default: \(0.1\))

originalSupport:

a logical indicating whether to use for minimum support the original definition of the support of a rule (lhs and rhs) instead of lhs support. Make sure to use ext = TRUE if originalSupport is set to FALSE (default: TRUE)

maxtime:

Time limit in seconds for checking subsets (default: 5 seconds)

Additional slots for Eclat in ECparameter:

tidLists:

a logical indicating whether to return also a list of supporting transactions (transaction IDs) (default: FALSE)

Methods

coerce

signature(from = "NULL", to = "APparameter")

coerce

signature(from = "list", to = "APparameter")

coerce

signature(from = "NULL", to = "ECparameter")

coerce

signature(from = "list", to = "ECparameter")

show

signature(object = "ASparameter")

References

Christian Borgelt (2004) Apriori --- Finding Association Rules/Hyperedges with the Apriori Algorithm. www.borgelt.net/apriori.html

See also

apriori, eclat