SunBai {arules}	R Documentation

The SunBai Data Set

Description

A small example database for weighted association rule mining provided as an object of class transactions.

Usage

data(SunBai)

Details

The data set contains the example database described in the paper by K. Sun and F.Bai for illustration of the concepts of weighted association rule mining. weight stored as transaction information denotes the transaction weights obtained using the HITS algorithm.

Source

K. Sun and F. Bai (2008). Mining Weighted Association Rules without Preassigned Weights. IEEE Transactions on Knowledge and Data Engineering, 4 (30), 489–495.

See Also

Class transactions, method transactionInfo, function hits.

Examples

data(SunBai)
summary(SunBai)
inspect(SunBai)

transactionInfo(SunBai)

[Package arules version 1.6-1 Index]