Calculate coverage for rules

Provides the generic function and the needed S4 method to calculate the coverage (support of the left-hand-side) of rules.

coverage(x, transactions = NULL, reuse = TRUE)

Arguments

x

the set of rules.

transactions

the data set used to generate 'x'. Only needed if the quality slot of 'x' does not contain support and confidence.

reuse

reuse support and confidence stored in 'x' or recompute from transactions?

Details

Coverage (also called cover or LHS-support) is the support of the left-hand-side of the rule, i.e., \(supp(X)\). It represents a measure of to how often the rule can be applied.

Coverage is quickly calculated from the rules quality measures (support and confidence) stored in the quality slot. If these values are not present, then the support of the LHS is counted using the data supplied in transactions.

Coverage is also one of the measures available via the function interestMeasures.

Value

A numeric vector of the same length as x containing the coverage values for the sets in x.

See also

interestMeasure, rules-class

Examples

data("Income")

## find and some rules (we only use 5 rules here) and calculate coverage
rules <- apriori(Income)[1:5]

#> Apriori
#> 
#> Parameter specification:
#>  confidence minval smax arem  aval originalSupport maxtime support minlen
#>         0.8    0.1    1 none FALSE            TRUE       5     0.1      1
#>  maxlen target   ext
#>      10  rules FALSE
#> 
#> Algorithmic control:
#>  filter tree heap memopt load sort verbose
#>     0.1 TRUE TRUE  FALSE TRUE    2    TRUE
#> 
#> Absolute minimum support count: 687 
#> 
#> set item appearances ...[0 item(s)] done [0.00s].
#> set transactions ...[50 item(s), 6876 transaction(s)] done [0.00s].
#> sorting and recoding items ... [30 item(s)] done [0.00s].
#> creating transaction tree ... done [0.00s].
#> checking subsets of size 1 2 3 4 5 6 7 8 done [0.06s].
#> writing ... [8664 rule(s)] done [0.00s].
#> creating S4 object  ... done [0.00s].

quality(rules) <- cbind(quality(rules), coverage = coverage(rules))

inspect(rules)

#>     lhs                                 rhs                               support confidence     lift  coverage
#> [1] {}                               => {language in home=english}      0.9128854  0.9128854 1.000000 1.0000000
#> [2] {occupation=clerical/service}    => {language in home=english}      0.1127109  0.9292566 1.017933 0.1212914
#> [3] {ethnic classification=hispanic} => {education=no college graduate} 0.1096568  0.8636884 1.224731 0.1269634
#> [4] {dual incomes=no}                => {marital status=married}        0.1400524  0.9441176 2.447871 0.1483421
#> [5] {dual incomes=no}                => {language in home=english}      0.1364165  0.9196078 1.007364 0.1483421