Mining Associations with Eclat

Mine frequent itemsets with the Eclat algorithm. This algorithm uses simple intersection operations for equivalence class clustering along with bottom-up lattice traversal.

eclat(data, parameter = NULL, control = NULL)

Arguments

data

object of class transactions or any data structure which can be coerced into transactions (e.g., binary matrix, data.frame).

parameter

object of class ECparameter or named list (default values are: support 0.1 and maxlen 5)

control

object of class ECcontrol or named list for algorithmic controls.

Details

Calls the C implementation of the Eclat algorithm by Christian Borgelt for mining frequent itemsets.

Note for contriol parameter tidLists=TRUE: Since storing transaction ID lists is very memory intensive, creating transaction ID lists only works for minimum support values which create a relatively small number of itemsets. See also supportingTransactions().

ruleInduction() can be used to generate rules from the found itemsets.

A weighted version of ECLAT is available as function weclat. This version can be used to perform weighted association rule mining (WARM).

Value

Returns an object of class itemsets.

References

Mohammed J. Zaki, Srinivasan Parthasarathy, Mitsunori Ogihara, and Wei Li. (1997) New algorithms for fast discovery of association rules. Technical Report 651, Computer Science Department, University of Rochester, Rochester, NY 14627.

Christian Borgelt (2003) Efficient Implementations of Apriori and Eclat. Workshop of Frequent Item Set Mining Implementations (FIMI 2003, Melbourne, FL, USA).

ECLAT Implementation: http://www.borgelt.net/eclat.html

See also

ECparameter-class, ECcontrol-class, transactions-class, itemsets-class, weclat, apriori, ruleInduction, supportingTransactions

Examples

data("Adult")
## Mine itemsets with minimum support of 0.1 and 5 or less items
itemsets <- eclat(Adult,
                parameter = list(supp = 0.1, maxlen = 5))

#> Eclat
#> 
#> parameter specification:
#>  tidLists support minlen maxlen            target   ext
#>     FALSE     0.1      1      5 frequent itemsets FALSE
#> 
#> algorithmic control:
#>  sparse sort verbose
#>       7   -2    TRUE
#> 
#> Absolute minimum support count: 4884 
#> 
#> create itemset ... 
#> set transactions ...[115 item(s), 48842 transaction(s)] done [0.03s].
#> sorting and recoding items ... [31 item(s)] done [0.01s].
#> creating bit matrix ... [31 row(s), 48842 column(s)] done [0.00s].
#> writing  ... [2143 set(s)] done [0.01s].
#> Creating S4 object  ... done [0.00s].

itemsets

#> set of 2143 itemsets 

## Create rules from the itemsets
rules <- ruleInduction(itemsets, Adult, confidence = .9)
rules

#> set of 2729 rules