image {arules}	R Documentation

Visual Inspection of Binary Incidence Matrices

Description

Provides the S4 methods image to generate level plots to visually inspect binary incidence matrices, i.e., objects based on itemMatrix (e.g., transactions, tidLists, items in itemsets or rhs/lhs in rules). These plots can be used to identify problems in a data set (e.g., recording problems with some transactions containing all items).

Usage

## S4 method for signature 'itemMatrix'
image(x, 
    xlab = "Items (Columns)",
    ylab = "Elements (Rows)", ...)

## S4 method for signature 'transactions'
image(x, 
    xlab = "Items (Columns)", 
    ylab = "Transactions (Rows)", ...)

## S4 method for signature 'tidLists'
image(x,
    xlab="Transactions (Columns)",
    ylab="Items/itemsets (Rows)", ...)

Arguments

x
the object (itemMatrix, transactions or tidLists).

xlab, ylab
labels for the plot.

...
further arguments passed on to image in package Matrix which in turn are passed on to levelplot in lattice.

Author(s)

Michael Hahsler

See Also

image (for dgTMatrix in Matrix), levelplot (in lattice), itemMatrix-class, transactions-class, tidLists-class

Examples

data("Epub")

## in this data set we can see that not all
## items were available from the beginning.
image(Epub[1:1000])

[Package arules version 1.6-1 Index]