inspect {arules}	R Documentation

Display Associations and Transactions in Readable Form

Description

Provides the generic function inspect and S4 methods to display associations and transactions plus additional information formatted for online inspection.

Usage

inspect(x, ...)

Arguments

x
a set of associations or transactions or an itemMatrix.

...
additional arguments can be used to customize the output: setStart, setEnd, itemSep and ruleSep. Items are printed only one per line in case the output lines get very long. This can also be directly controlled using linebreak.

Author(s)

Michael Hahsler and Kurt Hornik

See Also

itemMatrix-class, itemsets-class, rules-class, transactions-class

Examples

data("Adult")
rules <- apriori(Adult)
inspect(rules[1000])

inspect(rules[1000], ruleSep = "---->", itemSep = " + ", setStart = "", setEnd ="", 
  linebreak = FALSE)

[Package arules version 1.6-1 Index]