Calculating various additional interest measures

Provides the generic function interestMeasure and the needed S4 method to calculate various additional interest measures for existing sets of itemsets or rules.

interestMeasure(x, measure, transactions = NULL, reuse = TRUE, ...)

Arguments

x

a set of itemsets or rules.

measure

name or vector of names of the desired interest measures (see details for available measures). If measure is missing then all available measures are calculated.

transactions

the transaction data set used to mine the associations or a set of different transactions to calculate interest measures from (Note: you need to set reuse=FALSE in the later case).

reuse

logical indicating if information in quality slot should be reuse for calculating the measures. This speeds up the process significantly since only very little (or no) transaction counting is necessary if support, confidence and lift are already available. Use reuse=FALSE to force counting (might be very slow but is necessary if you use a differnt set of transactions than was used for mining).

…

further arguments for the measure calculation.

Details

For itemsets \(X\) the following measures are implemented:

"allConfidence" (Omiencinski, 2003)

Is defined on itemsets as the minimum confidence of all possible rule generated from the itemset.

Range: \([0, 1]\)

"crossSupportRatio", cross-support ratio (Xiong et al., 2003)

Defined on itemsets as the ratio of the support of the least frequent item to the support of the most frequent item, i.e., \(\frac{min(supp(x \in X))}{max(supp(x \in X))}\). Cross-support patterns have a ratio smaller than a set threshold. Normally many found patterns are cross-support patterns which contain frequent as well as rare items. Such patterns often tend to be spurious.

Range: \([0, 1]\)

"lift"

Probability (support) of the itemset over the product of the probabilities of all items in the itemset, i.e., \(\frac{supp(X)}{\prod_{x \in X} supp(X)}\). This is a measure of dependence similar to lift for rules.

Range: \([0, \infty]\) (1 indicated independence)

"support", supp (Agrawal et al., 1996)

Support is an estimate of \(P(X)\) a measure of generality of the itemset.

Range: \([0, 1]\)

For rules \(X \Rightarrow Y\) the following measures are implemented. In the following we use the notation \(supp(X \Rightarrow Y) = supp(X \cup Y)\) to indicate the support of the union of the itemsets \(X\) and \(Y\), i.e., the proportion of the transactions that contain both itemsets. We also use \(\overline{X}\) as the complement itemset to \(X\) with \(supp(\overline{X}) = 1 - supp(X)\), i.e., the proportion of transactions that do not contain \(X\).

"addedValue", added Value, AV, Pavillon index, centered confidence (Tan et al., 2002)

Defined as \(conf(X \Rightarrow Y) - supp(Y)\)

Range: \([-.5, 1]\)

"chiSquared", \(\chi^2\) (Liu et al., 1999)

The chi-squared statistic to test for independence between the lhs and rhs of the rule. The critical value of the chi-squared distribution with \(1\) degree of freedom (2x2 contingency table) at \(\alpha=0.05\) is \(3.84\); higher chi-squared values indicate that the lhs and the rhs are not independent. Note that the contingency table is likely to have cells with low expected values and that thus Fisher's Exact Test might be more appropriate (see below).

Called with significance=TRUE, the p-value of the test for independence is returned instead of the chi-squared statistic.

Range: \([0, \infty]\)

"certainty", certainty factor, CF, Loevinger (Berzal et al., 2002)

The certainty factor is a measure of variation of the probability that Y is in a transaction when only considering transactions with X. An inreasing CF means a decrease of the probability that Y is not in a transaction that X is in. Negative CFs have a similar interpretation.

Range: \([-1, 1]\) (0 indicates independence)

"collectiveStrength"

Collective strength (S).

Range: \([0, \infty]\)

"confidence", conf (Agrawal et al., 1996)

Rule confidence is an estimate of \(P(Y|X)\) calculated as \(\frac{supp(X \Rightarrow Y)}{supp(X)}\). Confidence is a measure of validity.

Range \([0, 1]\)

"conviction" (Brin et al. 1997)

Defined as \(\frac{supp(X)supp(\overline{Y})}{supp(X \cup \overline{Y})}\).

Range: \([.5, \infty]\) (\(1\) indicates unrelated items)

"cosine" (Tan et al., 2004)

Defined as \(\frac{supp(X \cup Y)}{\sqrt{(supp(X)supp(Y))}}\)

Range: \([0, 1]\)

"coverage", cover, LHS-support

Support of the left-hand-side of the rule, i.e., \(supp(X)\). A measure of to how often the rule can be applied.

Range: \([0, 1]\)

"confirmedConfidence", descriptive confirmed confidence (Kodratoff, 1999)

Confidence confirmed by its negative as \(conf(X \Rightarrow Y) - conf(X \Rightarrow \overline{Y})\).

Range: \([-1, 1]\)

"casualConfidence", casual confidence (Kodratoff, 1999)

Confidence reinforced by negatives given by \(\frac{1}{2}(conf(X \Rightarrow Y) + conf(\overline{Y} \Rightarrow \overline{X}))\).

Range: \([0, 1]\)

"casualSupport", casual support (Kodratoff, 1999)

Support improved by negatives given by \(supp(X \cup Y) - supp(\overline{X} \cup \overline{Y})\).

Range: \([-1, 1]\)

"counterexample", example and counterexample rate

\(\frac{supp(X \cup Y) - supp(X \cup \overline{Y})}{supp(X \cup Y)}\)

Range: \([0, 1]\)

"descriptiveConfirm", descriptive-confirm (Kodratoff, 1999)

Defined by \(supp(X \cup Y) - supp(X \cup \overline{Y})\).

Range: \([0, 1]\)

"doc", difference of confidence (Hofmann and Wilhelm, 2001)

Defined as \(conf(X \Rightarrow Y) - conf(\overline{X} \Rightarrow Y)\).

Range: \([-1, 1]\)

"fishersExactTest", Fisher's exact test (Hahsler and Hornik, 2007)

Statistical significance test used in the analysis of contingency tables where sample sizes are small. Returns the p-value. Note that it is equal to hyper-confidence with significance=TRUE.

Range: \([0, 1]\) (p-value scale)

"gini", Gini index (Tan et al., 2004)

Measures quadratic entropy.

Range: \([0, 1]\) (0 for independence)

"hyperLift" (Hahsler and Hornik, 2007)

Adaptation of the lift measure which is more robust for low counts. It is based on the idea that under independence the count \(c_{XY}\) of the transactions which contain all items in a rule \(X \Rightarrow Y\) follows a hypergeometric distribution (represented by the random variable \(C_{XY}\)) with the parameters given by the counts \(c_X\) and \(c_Y\).

Hyper-lift is defined as: $$\mathrm{hyperlift}(X \Rightarrow Y) = \frac{c_{XY}}{Q_{\delta}[C_{XY}]},$$

where \(Q_{\delta}[C_{XY}]\) is the quantile of the hypergeometric distribution given by \(\delta\). The quantile can be given as parameter d (default: d=0.99).

Range: \([0, \infty]\) (1 indicates independence)

"hyperConfidence" (Hahsler and Hornik, 2007)

Confidence level for observation of too high/low counts for rules \(X \Rightarrow Y\) using the hypergeometric model. Since the counts are drawn from a hypergeometric distribution (represented by the random variable \(C_{XY}\)) with known parameters given by the counts \(c_X\) and \(c_Y\), we can calculate a confidence interval for the observed counts \(c_{XY}\) stemming from the distribution. Hyper-confidence reports the confidence level (significance level if significance=TRUE is used) for

complements -

\(1 - P[C_{XY} >= c_{XY} | c_X, c_Y]\)

substitutes -

\(1 - P[C_{XY} < c_{XY} | c_X, c_Y]\).

A confidence level of, e.g., \(> 0.95\) indicates that there is only a 5% chance that the count for the rule was generated randomly.

Per default complementary effects are mined, substitutes can be found by using the parameter complements = FALSE.

Range: \([0, 1]\)

"imbalance", imbalance ratio, IR (Wu, Chen and Han, 2010)

IR is defined as \(\frac{|supp(X) - supp(Y)|}{supp(X) + supp(Y) - supp(X \Rightarrow Y))}\) gauges the degree of imbalance between two events that the lhs and the rhs are contained in a transaction. The ratio is close to 0 if the conditional probabilities are similar (i.e., very balanced) and close to 1 if they are very different.

Range: \([0, 1]\) (0 indicates a balanced rule)

"implicationIndex", implication index (Gras, 1996)

Defined as \(\sqrt{N} \frac{supp(X \cup \overline{Y}) - supp(X)supp(\overline{Y})}{\sqrt{supp(X)supp(\overline{Y})}}\). Represents a variation of the Lerman similarity.

Range: \([0, 1]\) (0 means independence)

"improvement" (Bayardo et al., 2000)

The improvement of a rule is the minimum difference between its confidence and the confidence of any more general rule (i.e., a rule with the same consequent but one or more items removed in the LHS). Defined as \(min_{X' \subset X}(conf(X \Rightarrow Y) - conf(X' \Rightarrow Y)\)

Range: \([0, 1]\)

"jaccard", Jaccard coefficient (Tan and Kumar, 2000)

Defined as \(\frac{supp(X \cup Y)}{supp(X) + supp(Y) - supp(X \cup Y)}\)

Range: \([-1, 1]\) (0 for independence)

"jMeasure", J-measure, J (Smyth and Goodman, 1991)

Measures cross entrophy.

Range: \([0, 1]\) (0 for independence)

"kappa" (Tan and Kumar, 2000)

Defined as \(\frac{supp(X \cup Y) + supp(\overline{X} \cup \overline{Y}) - supp(X)supp(Y) - supp(\overline{X})supp(\overline{Y})}{1- supp(X)supp(Y) - supp(\overline{X})supp(\overline{Y})}\)

Range: \([-1,1]\) (0 means independence)

"klosgen", Klosgen (Tan and Kumar, 2000)

Defined as \(\sqrt{supp(X \cup Y)} conf(X \Rightarrow Y) - supp(Y)\)

Range: \([-1, 1]\) (0 for independence)

"kulczynski" (Wu, Chen and Han, 2007; Kulczynski, 1927)

Calculate the null-invariant Kulczynski measure with a preference for skewed patterns.

Range: \([0, 1]\)

"lambda", Goodman-Kruskal \(\lambda\), predictive association (Tan and Kumar, 2000)

Range: \([0, 1]\)

"laplace", L (Tan and Kumar 2000)

Estimates confidence (decreases with lower support).

Range: \([0, 1]\)

"leastContradiction", least contradiction (Aze and Kodratoff, 2004

\(\frac{supp(X \cup Y) - supp(X \cup \overline{Y})}{supp(Y)}\).

Range: \([-1, 1]\)

"lerman", Lerman similarity (Lerman, 1981)

Defined as \(\sqrt{N} \frac{supp(X \cup Y) - supp(X)supp(Y)}{\sqrt{supp(X)supp(Y)}}\)

Range: \([0, 1]\)

"leverage", PS (Piatetsky-Shapiro 1991)

PS is defined as \(supp(X \Rightarrow Y) - supp(X)supp(Y)\). It measures the difference of X and Y appearing together in the data set and what would be expected if X and Y where statistically dependent. It can be interpreted as the gap to independence.

Range: \([-1, 1]\) (0 indicates intependence)

"lift", interest factor (Brin et al. 1997)

Lift quantifies dependence between X and Y by \(\frac{supp(X \cup Y)}{supp(X)supp(Y)}\).

Range: \([0, \infty]\) (1 means independence)

"mutualInformation", uncertainty, M (Tan et al., 2002)

Measures the information gain for Y provided by X.

Range: \([0, 1]\) (0 for independence)

"oddsRatio", odds ratio \(\alpha\) (Tan et al., 2004)

The odds of finding X in transactions which contain Y divided by the odds of finding X in transactions which do not contain Y.

Range: \([0, \infty]\) (\(1\) indicates that Y is not associated to X)

"phi", correlation coefficient \(\phi\) (Tan et al. 2004

Equivalent to Pearsons Product Moment Correlation Coefficient \(\rho\).

Range: \([-1, 1]\) (0 when X and Y are independent)

"ralambrodrainy", Ralambrodrainy Measure (Ralambrodrainy, 1991)

Range: \([0, 1]\)

"RLD", relative linkage disequilibrium (Kenett and Salini, 2008)

RLD evaluates the deviation of the support of the whole rule from the support expected under independence given the supports of the LHS and the RHS. The code was contributed by Silvia Salini.

Range: \([0, 1]\)

"sebag", Sebag measure (Sebag and Schoenauer, 1988)

Defined as \(\frac{supp(X \cup Y)}{supp(X \cup \overline{Y})}\)

Range: \([0, 1]\)

"support", supp (Agrawal et al., 1996)

Support is an estimate of \(P(X \cup Y)\) and measures the generality of the rule.

Range: \([0, 1]\)

"varyingLiaison", varying rates liaison (Bernard and Charron, 1996)

Defined as \(\frac{supp(X \cup Y)}{supp(X)supp(Y)}-1\). Is equivalent to \(lift(X \Rightarrow Y) -1\)

Range: \([-1, 1]\) (0 for independence)

"yuleQ", Yule's Q (Tan and Kumar, 2000)

Defined as \(\frac{\alpha-1}{\alpha+1}\) where \(\alpha\) is the odds ratio.

Range: \([-1, 1]\)

"yuleY", Yule's Y (Tan and Kumar, 2000)

Defined as \(\frac{\sqrt{\alpha}-1}{\sqrt{\alpha}+1}\) where \(\alpha\) is the odds ratio.

Range: \([-1, 1]\)

Value

If only one measure is used, the function returns a numeric vector containing the values of the interest measure for each association in the set of associations x.

If more than one measures are specified, the result is a data.frame containing the different measures for each association.

NA is returned for rules/itemsets for which a certain measure is not defined.

References

Agrawal, R., H Mannila, R Srikant, H Toivonen, AI Verkamo (1996). Fast Discovery of Association Rules. Advances in Knowledge Discovery and Data Mining 12 (1), 307--328.

Aze, J. and Y. Kodratoff (2004). Extraction de pepites de connaissances dans les donnees: Une nouvelle approche et une etude de sensibilite au bruit. In Mesures de Qualite pour la fouille de donnees. Revue des Nouvelles Technologies de l'Information, RNTI.

Bayardo, R. , R. Agrawal, and D. Gunopulos (2000). Constraint-based rule mining in large, dense databases. Data Mining and Knowledge Discovery, 4(2/3):217--240.

Berzal, Fernando, Ignacio Blanco, Daniel Sanchez and Maria-Amparo Vila (2002). Measuring the accuracy and interest of association rules: A new framework. Intelligent Data Analysis 6, 221--235.

Brin, Sergey, Rajeev Motwani, Jeffrey D. Ullman, and Shalom Tsur (1997). Dynamic itemset counting and implication rules for market basket data. In SIGMOD 1997, Proceedings ACM SIGMOD International Conference on Management of Data, pages 255--264, Tucson, Arizona, USA.

Hahsler, Michael and Kurt Hornik (2007). New probabilistic interest measures for association rules. Intelligent Data Analysis, 11(5):437--455.

Hofmann, Heike and Adalbert Wilhelm (2001). Visual comparison of association rules. Computational Statistics, 16(3):399--415.

Kenett, Ron and Silvia Salini (2008). Relative Linkage Disequilibrium: A New measure for association rules. In 8th Industrial Conference on Data Mining ICDM 2008, July 16--18, 2008, Leipzig/Germany.

Kodratoff, Y. (1999). Comparing Machine Learning and Knowledge Discovery in Databases: An Application to Knowledge Discovery in Texts. Lecture Notes on AI (LNAI) - Tutorial series.

Kulczynski, S. (1927). Die Pflanzenassoziationen der Pieninen. Bulletin International de l'Academie Polonaise des Sciences et des Lettres, Classe des Sciences Mathematiques et Naturelles B, 57--203.

Lerman, I.C. (1981). Classification et analyse ordinale des donnees. Paris.

Liu, Bing, Wynne Hsu, and Yiming Ma (1999). Pruning and summarizing the discovered associations. In KDD '99: Proceedings of the fifth ACM SIGKDD international conference on Knowledge discovery and data mining, pages 125--134. ACM Press, 1999.

Omiecinski, Edward R. (2003). Alternative interest measures for mining associations in databases. IEEE Transactions on Knowledge and Data Engineering, 15(1):57--69, Jan/Feb 2003.

Piatetsky-Shapiro, G. (1991). Discovery, analysis, and presentation of strong rules. In: Knowledge Discovery in Databases, pages 229--248.

Sebag, M. and M. Schoenauer (1988). Generation of rules with certainty and confidence factors from incomplete and incoherent learning bases. In Proceedings of the European Knowledge Acquisition Workshop (EKAW'88), Gesellschaft fuer Mathematik und Datenverarbeitung mbH, 28.1--28.20.

Smyth, Padhraic and Rodney M. Goodman (1991). Rule Induction Using Information Theory. Knowledge Discovery in Databases, 159--176.

Tan, Pang-Ning and Vipin Kumar (2000). Interestingness Measures for Association Patterns: A Perspective. TR 00-036, Department of Computer Science and Engineering University of Minnesota.

Tan, Pang-Ning, Vipin Kumar, and Jaideep Srivastava (2002). Selecting the right interestingness measure for association patterns. In Proceedings of the eighth ACM SIGKDD international conference on Knowledge discovery and data mining (KDD '02), ACM, 32--41.

Tan, Pang-Ning, Vipin Kumar, and Jaideep Srivastava (2004). Selecting the right objective measure for association analysis. Information Systems, 29(4):293--313.

Wu, Tianyi, Yuguo Chen, and Jiawei Han (2007). Association Mining in Large Databases: A Re-examination of Its Measures. In Proceedings of the 11th European conference on Principles and Practice of Knowledge Discovery in Databases (PKDD 2007), Springer-Verlag, Berlin, Heidelberg, 621--628.

Wu, T., Y. Chen, and J. Han (2010). Re-examination of interestingness measures in pattern mining: A unified framework. Data Mining and Knowledge Discovery, 2010.

Xiong, Hui, Pang-Ning Tan, and Vipin Kumar (2003). Mining strong affinity association patterns in data sets with skewed support distribution. In Bart Goethals and Mohammed J. Zaki, editors, Proceedings of the IEEE International Conference on Data Mining, November 19--22, 2003, Melbourne, Florida, pages 387--394.

See also

itemsets-class, rules-class

Examples

data("Income")
rules <- apriori(Income)

#> Apriori
#> 
#> Parameter specification:
#>  confidence minval smax arem  aval originalSupport maxtime support minlen
#>         0.8    0.1    1 none FALSE            TRUE       5     0.1      1
#>  maxlen target   ext
#>      10  rules FALSE
#> 
#> Algorithmic control:
#>  filter tree heap memopt load sort verbose
#>     0.1 TRUE TRUE  FALSE TRUE    2    TRUE
#> 
#> Absolute minimum support count: 687 
#> 
#> set item appearances ...[0 item(s)] done [0.00s].
#> set transactions ...[50 item(s), 6876 transaction(s)] done [0.00s].
#> sorting and recoding items ... [30 item(s)] done [0.00s].
#> creating transaction tree ... done [0.00s].
#> checking subsets of size 1 2 3 4 5 6 7 8 done [0.06s].
#> writing ... [8664 rule(s)] done [0.00s].
#> creating S4 object  ... done [0.00s].

## calculate a single measure and add it to the quality slot
quality(rules) <- cbind(quality(rules),
        hyperConfidence = interestMeasure(rules, measure = "hyperConfidence",
        transactions = Income))

inspect(head(rules, by = "hyperConfidence"))

#>     lhs                                 rhs                               support confidence     lift hyperConfidence
#> [1] {ethnic classification=hispanic} => {education=no college graduate} 0.1096568  0.8636884 1.224731               1
#> [2] {dual incomes=no}                => {marital status=married}        0.1400524  0.9441176 2.447871               1
#> [3] {occupation=student}             => {marital status=single}         0.1449971  0.8838652 2.160490               1
#> [4] {occupation=student}             => {age=14-34}                     0.1592496  0.9707447 1.658345               1
#> [5] {occupation=student}             => {dual incomes=not married}      0.1535777  0.9361702 1.564683               1
#> [6] {occupation=student}             => {income=$0-$40,000}             0.1381617  0.8421986 1.353027               1

## calculate several measures
m <- interestMeasure(rules, c("confidence", "oddsRatio", "leverage"),
        transactions = Income)
inspect(head(rules))

#>     lhs                                 rhs                               support confidence     lift hyperConfidence
#> [1] {}                               => {language in home=english}      0.9128854  0.9128854 1.000000       0.0000000
#> [2] {occupation=clerical/service}    => {language in home=english}      0.1127109  0.9292566 1.017933       0.9601859
#> [3] {ethnic classification=hispanic} => {education=no college graduate} 0.1096568  0.8636884 1.224731       1.0000000
#> [4] {dual incomes=no}                => {marital status=married}        0.1400524  0.9441176 2.447871       1.0000000
#> [5] {dual incomes=no}                => {language in home=english}      0.1364165  0.9196078 1.007364       0.7763615
#> [6] {occupation=student}             => {marital status=single}         0.1449971  0.8838652 2.160490       1.0000000

head(m)

#>   confidence oddsRatio     leverage
#> 1  0.9128854        NA 0.0000000000
#> 2  0.9292566  1.289208 0.0019856861
#> 3  0.8636884  2.952221 0.0201213950
#> 4  0.9441176 41.681686 0.0828384029
#> 5  0.9196078  1.107694 0.0009972213
#> 6  0.8838652 16.478646 0.0778840228

## calculate all available measures for the first 5 rules and show them as a 
## table with the measures as rows
t(interestMeasure(head(rules, 5), transactions = Income))

#>                              [,1]          [,2]          [,3]          [,4]
#> support              9.128854e-01  1.127109e-01  1.096568e-01  1.400524e-01
#> coverage             1.000000e+00  1.212914e-01  1.269634e-01  1.483421e-01
#> confidence           9.128854e-01  9.292566e-01  8.636884e-01  9.441176e-01
#> lift                 1.000000e+00  1.017933e+00  1.224731e+00  2.447871e+00
#> leverage             0.000000e+00  1.985686e-03  2.012140e-02  8.283840e-02
#> hyperLift            1.000000e+00  9.948652e-01  1.168992e+00  2.255269e+00
#> hyperConfidence      0.000000e+00  9.601859e-01  1.000000e+00  1.000000e+00
#> fishersExactTest     1.000000e+00  3.981412e-02  9.791397e-32  0.000000e+00
#> improvement                   Inf  1.637120e-02           Inf           Inf
#> chiSquared                     NA  3.198710e+00  1.208111e+02  1.576319e+03
#> cosine               9.554504e-01  3.387214e-01  3.664698e-01  5.855169e-01
#> conviction           1.000000e+00  1.231417e+00  2.162645e+00  1.099293e+01
#> gini                           NA  7.399053e-05  7.305254e-03  1.086335e-01
#> oddsRatio                      NA  1.289208e+00  2.952221e+00  4.168169e+01
#> phi                            NA  2.156848e-02  1.325518e-01  4.788000e-01
#> doc                            NA  1.863097e-02  1.815295e-01  6.556955e-01
#> RLD                            NA  1.879271e-01  5.376032e-01  9.090324e-01
#> imbalance            8.711460e-02  8.590593e-01  8.003221e-01  6.024363e-01
#> kulczynski           9.564427e-01  5.263616e-01  5.095922e-01  6.536199e-01
#> collectiveStrength   0.000000e+00  5.516834e+02  2.485516e+03  1.679296e+04
#> jaccard              9.128854e-01  1.223169e-01  1.517713e-01  3.554817e-01
#> kappa               -4.316065e+07 -8.854176e+06 -1.640126e+07 -2.744078e+07
#> mutualInformation              NA  8.289227e-04  2.622346e-02  2.934492e-01
#> lambda               0.000000e+00  0.000000e+00  0.000000e+00  3.416290e-01
#> jMeasure             0.000000e+00  2.172097e-04  8.880664e-03  1.055050e-01
#> laplace              9.127653e-01  9.282297e-01  8.628571e-01  9.432485e-01
#> certainty            0.000000e+00  1.879271e-01  5.376032e-01  9.090324e-01
#> addedValue           0.000000e+00  1.637120e-02  1.584819e-01  5.584283e-01
#> ralambrodrainy       8.711460e-02  8.580570e-03  1.730657e-02  8.289703e-03
#> descriptiveConfirm   8.257708e-01  1.041303e-01  9.235020e-02  1.317627e-01
#> confirmedConfidence  8.257708e-01  8.585132e-01  7.273769e-01  8.882353e-01
#> sebag                1.047913e+01  1.313559e+01  6.336134e+00  1.689474e+01
#> counterexample       9.045722e-01  9.238710e-01  8.421751e-01  9.408100e-01
#> casualSupport        1.738656e+00  1.017016e+00  7.975567e-01  5.174520e-01
#> casualConfidence     9.999735e-01  9.999883e-01  9.999766e-01  9.999887e-01
#> leastContradiction   9.045722e-01  1.140672e-01  1.309548e-01  3.416290e-01
#> centeredConfidence   0.000000e+00  1.637120e-02  1.584819e-01  5.584283e-01
#> varyingLiaison       0.000000e+00  1.793346e-02  2.247312e-01  1.447871e+00
#> yuleQ                          NA  1.263353e-01  4.939554e-01  9.531415e-01
#> yuleY                          NA  6.342171e-02  2.642197e-01  7.317645e-01
#> lerman               0.000000e+00  4.948292e-01  5.576076e+00  2.871764e+01
#> implicationIndex     0.000000e+00 -1.601836e+00 -8.624380e+00 -2.275482e+01
#>                              [,5]
#> support              1.364165e-01
#> coverage             1.483421e-01
#> confidence           9.196078e-01
#> lift                 1.007364e+00
#> leverage             9.972213e-04
#> hyperLift            9.873684e-01
#> hyperConfidence      7.763615e-01
#> fishersExactTest     2.236385e-01
#> improvement          6.722445e-03
#> chiSquared           6.805848e-01
#> cosine               3.707035e-01
#> conviction           1.083621e+00
#> gini                 1.574286e-05
#> oddsRatio            1.107694e+00
#> phi                  9.948857e-03
#> doc                  7.893362e-03
#> RLD                  7.716783e-02
#> imbalance            8.267023e-01
#> kulczynski           5.345211e-01
#> collectiveStrength   5.215988e+02
#> jaccard              1.475075e-01
#> kappa               -9.910284e+06
#> mutualInformation    1.706534e-04
#> lambda               0.000000e+00
#> jMeasure             4.316924e-05
#> laplace              9.187867e-01
#> certainty            7.716783e-02
#> addedValue           6.722445e-03
#> ralambrodrainy       1.192554e-02
#> descriptiveConfirm   1.244910e-01
#> confirmedConfidence  8.392157e-01
#> sebag                1.143902e+01
#> counterexample       9.125800e-01
#> casualSupport        1.037376e+00
#> casualConfidence     9.999864e-01
#> leastContradiction   1.363709e-01
#> centeredConfidence   6.722445e-03
#> varyingLiaison       7.363952e-03
#> yuleQ                5.109543e-02
#> yuleY                2.556441e-02
#> lerman               2.247083e-01
#> implicationIndex    -7.274143e-01

## calculate measures on a differnt set of transactions (I use a sample here)
## Note: reuse = TRUE (default) would just retun the stored support on the
##	data set used for mining
newTrans <- sample(Income, 100)
m2 <- interestMeasure(rules, "support", transactions = newTrans, reuse = FALSE)
head(m2)

#> [1] 0.92 0.11 0.13 0.14 0.13 0.13