is.significant {arules}	R Documentation

Find Significant Rules

Description

Provides the generic functions and the S4 method is.significant to find rules where the LHS and the RHS depend on each other. This uses Fisher's exact test and corrects for multiple comparisons.

Usage

is.significant(x, transactions, method = "fisher", 
  alpha = 0.01, adjust = "bonferroni") 

Arguments

x
a set of rules.

transactions
set of transactions used to mine the rules.

method
test to use. Options are "fisher", "chisq". Note that the contingency table is likely to have cells with low expected values and that thus Fisher's Exact Test might be more appropriate than the chi-squared test.

alpha
required significance level.

adjust
method to adjust for multiple comparisons. Options are "none", "bonferroni", "holm", "fdr", etc. (see p.adjust)

Value

returns a logical vector indicating which rules are significant.

Author(s)

Michael Hahsler

See Also

interestMeasure, p.adjust

Examples

data("Income")
rules <- apriori(Income, parameter = list(support = 0.5))
is.significant(rules, Income)

inspect(rules[is.significant(rules, Income)])

[Package arules version 1.6-1 Index]