itemFrequency {arules}	R Documentation

Getting Frequency/Support for Single Items

Description

Provides the generic function itemFrequency and S4 methods to get the frequency/support for all single items in an objects based on itemMatrix. For example, it is used to get the single item support from an object of class transactions without mining.

Usage

itemFrequency(x, ...)

## S4 method for signature 'itemMatrix'
itemFrequency(x, type, weighted = FALSE)

Arguments

x
an object.

...
further arguments are passed on.

type
a character string specifying if "relative" frequency/support or "absolute" frequency/support (item counts) is returned. (default: "relative").

weighted
should support be weighted by transactions weights stored as column "weight" in transactionInfo?

Value

itemFrequency returns a named numeric vector. Each element is the frequency/support of the corresponding item in object x. The items appear in the vector in the same order as in the binary matrix in x.

Author(s)

Michael Hahsler

See Also

itemFrequencyPlot, itemMatrix-class, transactions-class

Examples

data("Adult")
itemFrequency(Adult, type = "relative")

[Package arules version 1.6-1 Index]