Merging (adding) items

Provides the generic function merge and the S4 methods for itemMatrix and transactions. The methods are used to add new items to existing data.

merge(x, y, ...)

Arguments

x

an object of class itemMatrix or transactions.

y

an object of the same class as x (or something which can be coerced to that class).

...

further arguments; unused.

Value

Returns a new object of the same class as x with the items in y added.

See also

transactions-class, itemMatrix-class, addComplement

Examples

data("Groceries")

## create a random item as a matrix
randomItem <- sample(c(TRUE, FALSE), size=length(Groceries),replace=TRUE)
randomItem <- as.matrix(randomItem)
colnames(randomItem) <- "random item"
head(randomItem, 3)

#>      random item
#> [1,]       FALSE
#> [2,]       FALSE
#> [3,]        TRUE

## add the random item to Groceries
g2 <- merge(Groceries, randomItem)
nitems(Groceries)

#> [1] 169

nitems(g2)

#> [1] 170

inspect(head(g2,3))

#>     items                
#> [1] {citrus fruit,       
#>      semi-finished bread,
#>      margarine,          
#>      ready soups}        
#> [2] {tropical fruit,     
#>      yogurt,             
#>      coffee}             
#> [3] {whole milk,         
#>      random item}