merge {arules}	R Documentation

Adding Items to Data

Description

Provides the generic function merge and the S4 methods for itemMatrix and transactions. The methods are used to add new items to existing data.

Usage

merge(x, y, ...)

Arguments

x
an object of class itemMatrix or transactions.

y
an object of the same class as x (or something which can be coerced to that class).

...
further arguments; unused.

Value

Returns a new object of the same class as x with the items in y added.

Author(s)

Michael Hahsler

See Also

transactions-class, itemMatrix-class, addComplement

Examples

data("Groceries")

## create a random item as a matrix
randomItem <- sample(c(TRUE, FALSE), size=length(Groceries),replace=TRUE)
randomItem <- as.matrix(randomItem)
colnames(randomItem) <- "random item"
head(randomItem, 3)

## add the random item to Groceries
g2 <- merge(Groceries, randomItem)
nitems(Groceries)
nitems(g2)
inspect(head(g2,3))

[Package arules version 1.6-1 Index]