proximity-classes {arules}	R Documentation

Classes dist, ar\_cross\_dissimilarity and ar\_similarity — Proximity Matrices

Description

Simple classes to represent proximity matrices. For compatibility with clustering functions in R, we represent dissimilarities as the S3 class dist. For cross-dissimilarities and similarities, we provide the S4 classes ar_cross_dissimilarities and ar_similarities.

Objects from the Class

dist objects are the result of calling the method dissimilarity with one argument or any R function returning a S3 dist object.

ar_cross_dissimilarity objects are the result of calling the method dissimilarity with two arguments, by calls of the form new("similarity", ...), or by coercion from matrix.

ar_similarity objects are the result of calling the method affinity, by calls of the form new("similarity", ...), or by coercion from matrix.

Slots

The S4 classes have a method slot which contains the type of measure used for calculation.

Author(s)

Michael Hahsler

See Also

dist (in package stats), dissimilarity, affinity.

[Package arules version 1.6-1 Index]