Rule Induction for a Set of Itemsets

Provides the generic function and the needed S4 method to induce all rules which can be generated by the given itemsets from a transactions data set.

ruleInduction(x, ...)
# S4 method for itemsets
ruleInduction(x, transactions, confidence = 0.8,
    control = NULL)

Arguments

x

the set of itemsets from which rules will be induced.

…

further arguments.

transactions

the transaction data set used to mine the itemsets. Can be omitted if x contains a lattice (complete set) of frequent itemsets together with their support counts.

confidence

a numeric value giving the minimum confidence for the rules.

control

a named list with elements method indicating the method ("apriori" or "ptree"), and the logical arguments reduce and verbose to indicate if unused items are removed and if the output should be verbose. Currently, "ptree" is the default method.

Details

If in control method = "apriori" is used, a very simple rule induction method is used. All rules are mined from the transactions data set using Apriori with the minimal support found in itemsets. And in a second step all rules which do not stem from one of the itemsets are removed. This procedure will be in many cases very slow (e.g., for itemsets with many elements or very low support).

If in control method = "ptree" is used, the transactions are counted into a prefix tree and then the rules are selectively generated using the counts in the tree. This is usually faster than the above approach.

If in control reduce = TRUE is used, unused items are removed from the data before creating rules. This might be slower for large transaction data sets. However, for method = "ptree" this is highly recommended as the items are further reordered to reduce the counting time.

If argument transactions is missing it is assumed that x contains a lattice (complete set) of frequent itemsets together with their support counts. Then rules can be induced directly without support counting. This approach is very fast.

For transactions, a set different to the data used for creating the original itemsets can be used, however, the new set has to conform in terms of items and their order.

Value

An object of class rules.

References

Michael Hahsler, Christian Buchta, and Kurt Hornik. Selective association rule generation. Computational Statistics, 23(2):303-315, April 2008

See also

itemsets-class, rules-class transactions-class

Examples

data("Adult")

## find all closed frequent itemsets
closed <- apriori(Adult,
        parameter = list(target = "closed", support = 0.4))

#> Apriori
#> 
#> Parameter specification:
#>  confidence minval smax arem  aval originalSupport maxtime support minlen
#>          NA    0.1    1 none FALSE            TRUE       5     0.4      1
#>  maxlen                   target   ext
#>      10 closed frequent itemsets FALSE
#> 
#> Algorithmic control:
#>  filter tree heap memopt load sort verbose
#>     0.1 TRUE TRUE  FALSE TRUE    2    TRUE
#> 
#> Absolute minimum support count: 19536 
#> 
#> set item appearances ...[0 item(s)] done [0.00s].
#> set transactions ...[115 item(s), 48842 transaction(s)] done [0.03s].
#> sorting and recoding items ... [11 item(s)] done [0.01s].
#> creating transaction tree ... done [0.03s].
#> checking subsets of size 1 2 3 4 5 done [0.00s].
#> filtering closed item sets ... done [0.00s].
#> writing ... [99 set(s)] done [0.00s].
#> creating S4 object  ... done [0.01s].

closed

#> set of 99 itemsets 

## rule induction
rules <- ruleInduction(closed, Adult, control = list(verbose = TRUE))

#> ruleInduction: using method ptree 
#> preprocessing done [0s].
#> preparing ... 593 itemsets, created 203 (0.20) nodes [0.00s]
#> counting ... 48842 transactions, processed 9285411 (0.31) nodes [0.07s]
#> writing ... 247 rules, processed 1433 (0.70) nodes [0.00s]
#> searching done [0.092s].
#> postprocessing done [0s].

summary(rules)

#> set of 165 rules
#> 
#> rule length distribution (lhs + rhs):sizes
#>  2  3  4  5 
#> 40 74 43  8 
#> 
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   2.000   3.000   3.000   3.115   4.000   5.000 
#> 
#> summary of quality measures:
#>     support         confidence          lift           itemset     
#>  Min.   :0.4013   Min.   :0.8209   Min.   :0.9594   Min.   :12.00  
#>  1st Qu.:0.4341   1st Qu.:0.8871   1st Qu.:0.9912   1st Qu.:47.00  
#>  Median :0.4994   Median :0.9129   Median :0.9984   Median :70.00  
#>  Mean   :0.5333   Mean   :0.9121   Mean   :1.0409   Mean   :65.33  
#>  3rd Qu.:0.5697   3rd Qu.:0.9471   3rd Qu.:1.0169   3rd Qu.:87.00  
#>  Max.   :0.8707   Max.   :0.9999   Max.   :2.4529   Max.   :99.00  
#> 
#> mining info:
#>   data ntransactions support confidence
#>  Adult         48842     0.4        0.8

## inspect the resulting rules
inspect(head(rules, by = "lift"))

#>     lhs                                    rhs                                   support confidence     lift itemset
#> [1] {marital-status=Married-civ-spouse,                                                                             
#>      sex=Male}                          => {relationship=Husband}              0.4034028  0.9901503 2.452877      47
#> [2] {relationship=Husband}              => {marital-status=Married-civ-spouse} 0.4034233  0.9993914 2.181164      12
#> [3] {marital-status=Married-civ-spouse} => {relationship=Husband}              0.4034233  0.8804683 2.181164      12
#> [4] {relationship=Husband,                                                                                          
#>      sex=Male}                          => {marital-status=Married-civ-spouse} 0.4034028  0.9993913 2.181164      47
#> [5] {relationship=Husband}              => {sex=Male}                          0.4036485  0.9999493 1.495851      13
#> [6] {marital-status=Married-civ-spouse,                                                                             
#>      relationship=Husband}              => {sex=Male}                          0.4034028  0.9999492 1.495851      47

## use lattice of frequent itemsets
ec  <- eclat(Adult, parameter = list(support = 0.4))

#> Eclat
#> 
#> parameter specification:
#>  tidLists support minlen maxlen            target   ext
#>     FALSE     0.4      1     10 frequent itemsets FALSE
#> 
#> algorithmic control:
#>  sparse sort verbose
#>       7   -2    TRUE
#> 
#> Absolute minimum support count: 19536 
#> 
#> create itemset ... 
#> set transactions ...[115 item(s), 48842 transaction(s)] done [0.03s].
#> sorting and recoding items ... [11 item(s)] done [0.01s].
#> creating bit matrix ... [11 row(s), 48842 column(s)] done [0.00s].
#> writing  ... [99 set(s)] done [0.00s].
#> Creating S4 object  ... done [0.00s].

rec <- ruleInduction(ec)
rec

#> set of 165 rules 

inspect(head(rec))

#>     lhs                                    rhs                                   support confidence     lift
#> [1] {relationship=Husband,                                                                                  
#>      sex=Male}                          => {marital-status=Married-civ-spouse} 0.4034028  0.9993913 2.181164
#> [2] {marital-status=Married-civ-spouse,                                                                     
#>      sex=Male}                          => {relationship=Husband}              0.4034028  0.9901503 2.452877
#> [3] {marital-status=Married-civ-spouse,                                                                     
#>      relationship=Husband}              => {sex=Male}                          0.4034028  0.9999492 1.495851
#> [4] {relationship=Husband}              => {sex=Male}                          0.4036485  0.9999493 1.495851
#> [5] {relationship=Husband}              => {marital-status=Married-civ-spouse} 0.4034233  0.9993914 2.181164
#> [6] {marital-status=Married-civ-spouse} => {relationship=Husband}              0.4034233  0.8804683 2.181164