Class rules --- A Set of Rules

The rules class represents a set of rules.

Note that the class can also represent a multiset of rules with duplicated elements. Duplicated elements can be removed with unique.

Objects from the Class

Objects are the result of calling the function apriori. Objects can also be created by calls of the form new("rules", ...).

Slots

lhs:

Object of class itemMatrix; the left-hand-sides of the rules (antecedents)

rhs:

Object of class itemMatrix; the right-hand-sides of the rules (consequents)

quality:

a data.frame

Extends

Class associations, directly.

Methods

% \item{[}{\code{signature(x = "rules")}; % extracts a subset of rules and the associated quality measures}
coerce

signature(from = "rules", to = "data.frame"); represents the set of rules as a data.frame

itemInfo

signature(object = "rules"); returns the whole item information data frame including item labels

itemLabels

signature(object = "rules"); returns the item labels used to encode the rules

items

signature(x = "rules"); returns for each rule the union of the items in the lhs and rhs (i.e., the itemsets which generated the rule) as an itemMatrix

generatingItemsets

signature(x = "rules"); returns a collection of the itemsets which generated the rules (one itemset for each rule). Note that the collection can be a multiset and contain duplicated elements. Use unique to remove duplicates and obtain a proper set.

labels

signature(object = "rules"); returns labels for the rules ("lhs => rhs") as a character vector. The representation can be customized using the additional parameter ruleSep and parameters for label defined in itemMatrix

itemLabels

signature(object = "rules"); returns the item labels as a character vector. The index for each label is the column index of the item in the binary matrix.
% \item{length}{\code{signature(x = "rules")}; % returns the number of rules stored in the the set}
lhs

signature(x = "rules"); returns the itemMatrix representing the left-hand-side of the rules (antecedents)

lhs<-

signature(x = "rules"); replaces the itemMatrix representing the left-hand-side of the rules (antecedents)

rhs

signature(x = "rules"); returns the itemMatrix representing the right-hand-side of the rules (consequents)

rhs<-

signature(x = "rules"); replaces the itemMatrix representing the right-hand-side of the rules (consequents)
%\item{subset}{\code{signature(x = "rules")}; % selects a subset using restrictions on the quality measures or on % the items present in the rules (see examples).}
summary

signature(object = "rules")

See also

[-methods, apriori, c, duplicated, inspect, length, match, sets, size, subset, associations-class, itemMatrix-class,

Examples

data("Adult")

## Mine rules.
rules <- apriori(Adult, parameter = list(support = 0.4))

#> Apriori
#> 
#> Parameter specification:
#>  confidence minval smax arem  aval originalSupport maxtime support minlen
#>         0.8    0.1    1 none FALSE            TRUE       5     0.4      1
#>  maxlen target   ext
#>      10  rules FALSE
#> 
#> Algorithmic control:
#>  filter tree heap memopt load sort verbose
#>     0.1 TRUE TRUE  FALSE TRUE    2    TRUE
#> 
#> Absolute minimum support count: 19536 
#> 
#> set item appearances ...[0 item(s)] done [0.00s].
#> set transactions ...[115 item(s), 48842 transaction(s)] done [0.03s].
#> sorting and recoding items ... [11 item(s)] done [0.01s].
#> creating transaction tree ... done [0.03s].
#> checking subsets of size 1 2 3 4 5 done [0.00s].
#> writing ... [169 rule(s)] done [0.00s].
#> creating S4 object  ... done [0.01s].

## Select a subset of rules using partial matching on the items 
## in the right-hand-side and a quality measure
rules.sub <- subset(rules, subset = rhs %pin% "sex" & lift > 1.3)

## Display the top 3 support rules
inspect(head(rules.sub, n = 3, by = "support"))

#>     lhs                                    rhs          support confidence     lift
#> [1] {marital-status=Married-civ-spouse} => {sex=Male} 0.4074157  0.8891818 1.330151
#> [2] {relationship=Husband}              => {sex=Male} 0.4036485  0.9999493 1.495851
#> [3] {marital-status=Married-civ-spouse,                                            
#>      relationship=Husband}              => {sex=Male} 0.4034028  0.9999492 1.495851

## Display the first 3 rules
inspect(rules.sub[1:3])

#>     lhs                                    rhs          support confidence     lift
#> [1] {relationship=Husband}              => {sex=Male} 0.4036485  0.9999493 1.495851
#> [2] {marital-status=Married-civ-spouse} => {sex=Male} 0.4074157  0.8891818 1.330151
#> [3] {marital-status=Married-civ-spouse,                                            
#>      relationship=Husband}              => {sex=Male} 0.4034028  0.9999492 1.495851

## Get labels for the first 3 rules
labels(rules.sub[1:3])

#> [1] "{relationship=Husband} => {sex=Male}"                                  
#> [2] "{marital-status=Married-civ-spouse} => {sex=Male}"                     
#> [3] "{marital-status=Married-civ-spouse,relationship=Husband} => {sex=Male}"

labels(rules.sub[1:3], itemSep = " + ", setStart = "", setEnd="",
  ruleSep = " ---> ")

#> [1] "relationship=Husband ---> sex=Male"                                    
#> [2] "marital-status=Married-civ-spouse ---> sex=Male"                       
#> [3] "marital-status=Married-civ-spouse + relationship=Husband ---> sex=Male"