sample {arules}	R Documentation

Random Samples and Permutations

Description

Provides the generic function sample and the S4 method to take a sample of the specified size from the elements of x using either with or without replacement. sample can be used to sample from a set of transactions or associations.

Usage

sample(x, size, replace = FALSE, prob = NULL, ...)

Arguments

x
object to be sampled from (a set of associations or transactions).

size
sample size.

replace
a logical. Sample with replacement?

prob
a numeric vector of probability weights.

...
further arguments.

Value

An object of the same class as x.

Author(s)

Michael Hahsler

See Also

associations-class, transactions-class, itemMatrix-class.

Examples

data("Adult")

## sample with replacement
s <- sample(Adult, 500, replace = TRUE)
s

[Package arules version 1.6-1 Index]