tidLists-class {arules}	R Documentation

Class tidLists — Transaction ID Lists for Items/Itemsets

Description

Transaction ID lists contains a set of lists. Each list is associated with an item/itemset and stores the IDs of the transactions which support the item/itemset. tidLists uses the class ngCMatrix to efficiently store the transaction ID lists as a sparse matrix. Each column in the matrix represents one transaction ID list.

tidLists can be used for different purposes. For some operations (e.g., support counting) it is efficient to coerce a transactions database into tidLists where each list contains the transaction IDs for an item (and the support is given by the length of the list).

The implementation of the Eclat mining algorithm (which uses transaction ID list intersection) can also produce transaction ID lists for the found itemsets as part of the returned itemsets object. These lists can then be used for further computation.

Objects from the Class

Objects are created by Eclat if the eclat function is called with tidLists = TRUE in the ECparameter object, and returned as part of the mined itemsets. Objects can also be created by coercion from an object of class transactions or by calls of the form new("tidLists", ...).

Slots

data:

object of class ngCMatrix.

itemInfo:

a data.frame to store item/itemset labels (see itemMatrix class).

transactionInfo:

a data.frame with vectors of the same length as the number of transactions. Each vector can hold additional information e.g., store transaction IDs or user IDs for each transaction.

Methods

coerce

signature(from = "tidLists", to = "ngCMatrix"); access the sparse matrix representation. In the ngCMatrix each column represents the transaction IDs for one item/itemset.

coerce

signature(from = "tidLists", to = "list")

coerce

signature(from = "list", to = "tidLists")

coerce

signature(from = "tidLists", to = "matrix")

coerce

signature(from = "tidLists", to = "itemMatrix")

coerce

signature(from = "tidLists", to = "transactions")

coerce

signature(from = "itemMatrix", to = "tidLists"); this also coerces from transactions.

coerce

signature(from = "transactions", to = "tidLists")

c

signature(x = "tidLists"); combine.

dim

signature(x = "tidLists"); returns the dimensions of the sparse Matrix representing the tidLists.

dimnames, rownames, colnames

signature(x = "transactions"); returns row (items/itemsets) and column (transactionIDs if available) names.

labels

signature(x = "transactions"); returns the labels for the itemsets in each transaction (see itemMatrix).

inspect

inspect the transaction ID lists.

itemInfo

returns the slot itemInfo.

itemLabels

signature(object = "tidLists"); returns the item labels as a character vector.

labels

signature(x = "transactions"); returns the labels (transaction IDs).

show

signature(object = "tidLists")

summary

signature(object = "tidLists")

transactionInfo

signature(x = "transactions"): returns the slot transactionInfo.

Author(s)

Michael Hahsler

See Also

[-methods, LIST, eclat, image, length, size, ngCMatrix(in Matrix), itemMatrix-class, itemsets-class, transactions-class

Examples

## Create transaction data set.
data <- list(
  c("a","b","c"),
  c("a","b"),
  c("a","b","d"),
  c("b","e"),
  c("b","c","e"),
  c("a","d","e"),
  c("a","c"),
  c("a","b","d"),
  c("c","e"),
  c("a","b","d","e")
  )
data <- as(data, "transactions")
data

## convert transactions to transaction ID lists
tl <- as(data, "tidLists")
tl 

inspect(tl)
dim(tl)
dimnames(tl)

## inspect visually
image(tl)

## mine itemsets with transaction ID lists
f <- eclat(data, parameter = list(support = 0, tidLists = TRUE))
tl2 <- tidLists(f)
inspect(tl2)

[Package arules version 1.6-1 Index]