write {arules}	R Documentation

Write Transactions or Associations to a File

Description

Provides the generic function write and the S4 methods to write transactions or associations (itemsets, rules) to a file.

Usage

write(x, file = "",...)
## S4 method for signature 'transactions'
write(x, file="", format = c("basket", "single"),
                sep=" ", quote=TRUE, ...)
## S4 method for signature 'associations'
write(x, file="", sep=" ", quote=TRUE, ...)

Arguments

x
the transactions or associations (rules, itemsets, etc.) object.

file
either a character string naming a file or a connection open for writing. '""' indicates output to the console.

format
format to write transactions.

sep
the field separator string. Values within each row of x are separated by this string. Use quote=TRUE and sep="," for saving data as in csv format.

quote
a logical value. Quote fields?

...
further arguments passed on to write.table or write. Use fileEncoding to set the encoding used for writing the file.

Details

For associations (rules and itemsets) write first uses coercion to data.frame to obtain a printable form of x and then uses write.table to write the data to disk.

Transactions can be saved in basket (one line per transaction) or in single (one line per item) format.

Note: To save and load associations in compact form, use save and load from the base package. Alternatively, association can be written to disk in PMML (Predictive Model Markup Language) via write.PMML. This requiresvpackages pmml.

Author(s)

Michael Hahsler

See Also

read.transactions for reading transactions from a file, read.PMML and write.PMML for reading/writing associations in PMML format, write.table (in base), transactions-class, associations-class

Examples

data("Epub")

## write the formated transactions to screen (basket format)
write(head(Epub))

## write the formated transactions to screen (single format)
write(head(Epub), format="single")

## write the formated result to file in CSV format
write(Epub, file = "data.csv", format="single", sep = ",")

## write rules in CSV format
rules <- apriori(Epub, parameter=list(support=0.0005, conf=0.8))
write(rules, file = "data.csv", sep = ",")

unlink("data.csv") # tidy up

[Package arules version 1.6-1 Index]